refactor user resolvers into own files
